+++ client/src/app/components/register/organization/organization.tsx
@@ -21,8 +21,8 @@ import {TipContent, TipVariant} from '../../../ui';
 import {NextStep} from '../common/buttons';
 import {PartViewProps} from '../common/part-view-props';
 
-import {CreateNewOrganization} from './@create-new-organization';
-import {JoinExistOrganization} from './@join-exist-organization';
+import {CreateNewOrganization} from './create-new-organization';
+import {JoinExistOrganization} from './join-exist-organization';
 
 const Wrapper = styled.div``;
